Moved export classes out of contracts/ subfolder to match codebase structure, fixed spacing and declare placement (#56)
