+++ classes/export/datatypes/float_field.php
@@ -14,10 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export\datatypes;
-
-use Exception;
-
 /**
  * Handles floating-point number validation and export for custom certificate fields.
  *
@@ -29,6 +25,13 @@ use Exception;
  * @author     Konrad Ebel <konrad.ebel@oncampus.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export\datatypes;
+
+use Exception;
+
 class float_field implements i_field {
     /**
      * @var float|null Minimum allowed value, or null if no lower bound.
