+++ classes/export/datatypes/format_error.php
@@ -14,10 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export\datatypes;
-
-use Exception;
-
 /**
  * Represents a fatal error during certificate export/import formatting.
  *
@@ -29,5 +25,12 @@ use Exception;
  * @author     Konrad Ebel <konrad.ebel@oncampus.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export\datatypes;
+
+use Exception;
+
 class format_error extends Exception {
 }
