+++ classes/export/datatypes/int_field.php
@@ -14,10 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export\datatypes;
-
-use Exception;
-
 /**
  * Handles integer-specific import functionality for certificate subplugin fields.
  *
@@ -26,6 +22,13 @@ use Exception;
  * @author     Konrad Ebel <konrad.ebel@oncampus.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export\datatypes;
+
+use Exception;
+
 class int_field extends float_field {
     /**
      * Imports and validates a numeric value, casting the result to an integer.
