+++ classes/export/datatypes/unimported_field.php
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export\datatypes;
-
 /**
  * Represents a field that cannot be imported in custom certificate exports.
  *
@@ -26,6 +24,12 @@ namespace mod_customcert\export\datatypes;
  * @author     Konrad Ebel <konrad.ebel@oncampus.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export\datatypes;
+
+
 class unimported_field implements i_field {
     /**
      * Throws an exception to indicate that this field is not importable.
