+++ classes/export/import_exception.php
@@ -14,10 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export\contracts;
-
-use moodle_exception;
-
 /**
  * Fatal exception during template import.
  *
@@ -28,5 +24,12 @@ use moodle_exception;
  * @copyright  2025, oncampus GmbH
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export;
+
+use moodle_exception;
+
 class import_exception extends moodle_exception {
 }
