+++ classes/export/template.php
@@ -14,13 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export;
-
-use core\clock;
-use core\di;
-use mod_customcert\export\contracts\import_exception;
-use moodle_database;
-
 /**
  * Handles the import and export of custom certificate templates.
  *
@@ -33,6 +26,16 @@ use moodle_database;
  * @copyright  2025, oncampus GmbH
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export;
+
+use core\clock;
+use core\di;
+use mod_customcert\export\import_exception;
+use moodle_database;
+
 class template {
     /**
      * @var clock Clock service for generating consistent timestamps.
