+++ classes/export/template_appendix_manager.php
@@ -14,13 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export;
-
-use coding_exception;
-use Exception;
-use mod_customcert\export\contracts\i_template_appendix_manager;
-use stored_file;
-
 /**
  * Manages the export and import of appendix files for custom certificate templates.
  *
@@ -33,6 +26,16 @@ use stored_file;
  * @copyright  2025, oncampus GmbH
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export;
+
+use coding_exception;
+use Exception;
+use mod_customcert\export\i_template_appendix_manager;
+use stored_file;
+
 class template_appendix_manager implements i_template_appendix_manager {
     /** @var array<string, stored_file> In-memory index of imported files, mapped by content hash. */
     private array $index = [];
