+++ classes/export/template_file_manager.php
@@ -14,13 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_customcert\export;
-
-use coding_exception;
-use mod_customcert\export\contracts\import_exception;
-use mod_customcert\export\contracts\i_template_file_manager;
-use mod_customcert\export\contracts\i_template_appendix_manager;
-
 /**
  * Manages the export and import of custom certificate template files.
  *
@@ -34,6 +27,16 @@ use mod_customcert\export\contracts\i_template_appendix_manager;
  * @copyright  2025, oncampus GmbH
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace mod_customcert\export;
+
+use coding_exception;
+use mod_customcert\export\import_exception;
+use mod_customcert\export\i_template_file_manager;
+use mod_customcert\export\i_template_appendix_manager;
+
 class template_file_manager implements i_template_file_manager {
     /** @var i_template_appendix_manager The manager for appendix file operations. */
     private readonly i_template_appendix_manager $filemng;
