+++ cli/import_template.php
@@ -25,7 +25,7 @@
 
 use core\di;
 use core\session\manager;
-use mod_customcert\export\contracts\i_template_file_manager;
+use mod_customcert\export\i_template_file_manager;
 
 define('WORKING_DIR', getcwd());
 define('CLI_SCRIPT', true);
