+++ element/code/classes/exporter.php
@@ -14,10 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace customcertelement_code;
-
-use mod_customcert\export\contracts\subplugin_text_exportable;
-
 /**
  * Empty custom data exporter.
  *
@@ -26,5 +22,12 @@ use mod_customcert\export\contracts\subplugin_text_exportable;
  * @copyright  2025, oncampus GmbH
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace customcertelement_code;
+
+use mod_customcert\export\subplugin_text_exportable;
+
 class exporter extends subplugin_text_exportable {
 }
