+++ element/image/classes/exporter.php
@@ -14,13 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace customcertelement_image;
-
-use mod_customcert\export\datatypes\file_field;
-use mod_customcert\export\datatypes\float_field;
-use mod_customcert\export\contracts\subplugin_exportable;
-use mod_customcert\export\datatypes\i_field;
-
 /**
  * Handles import and export of image elements for custom certificates.
  *
@@ -29,6 +22,16 @@ use mod_customcert\export\datatypes\i_field;
  * @copyright  2025, oncampus GmbH
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+declare(strict_types=1);
+
+namespace customcertelement_image;
+
+use mod_customcert\export\datatypes\file_field;
+use mod_customcert\export\datatypes\float_field;
+use mod_customcert\export\subplugin_exportable;
+use mod_customcert\export\datatypes\i_field;
+
 class exporter extends subplugin_exportable {
     /**
      * Defines the custom data fields
