+++ pages/export.php
@@ -31,7 +31,7 @@ require_once(__DIR__ . '/../../../config.php');
 require_once($CFG->libdir . '/filelib.php');
 
 use core\di;
-use mod_customcert\export\contracts\i_template_file_manager;
+use mod_customcert\export\i_template_file_manager;
 
 require_login();
