+++ pages/import.php
@@ -31,9 +31,9 @@ require_once(__DIR__ . '/../../../config.php');
 
 use core\di;
 use core\notification;
-use mod_customcert\export\contracts\i_template_import_logger;
-use mod_customcert\export\contracts\i_template_file_manager;
-use mod_customcert\export\contracts\import_form;
+use mod_customcert\export\i_template_import_logger;
+use mod_customcert\export\i_template_file_manager;
+use mod_customcert\export\import_form;
 
 require_login();
